+++ app/src/main/java/msi/paria/trainingapp/pages/main/state/MainPageState.kt
@@ -1,3 +1,3 @@
-package msi.paria.trainingapp
+package msi.paria.trainingapp.pages.main.state
 
 data class MainPageState(var email:String = "" , var isEmailCorrect: Boolean = false)
